Separate ineligible claim feature specs
A bit of re-organising of the specs to make them easier to work with –
the main tslr_claim_spec.rb file was getting large and unweildy.
