+++ spec/features/tslr_claim_spec.rb
@@ -106,63 +106,6 @@ RSpec.feature "Teacher Student Loan Repayments claims" do
     expect(page).to have_text(claim.email_address)
   end
 
-  scenario "Teacher now works for a different school" do
-    claim = start_tslr_claim
-    choose_qts_year
-    choose_school schools(:penistone_grammar_school)
-
-    choose_still_teaching "Yes, at another school"
-
-    expect(claim.reload.employment_status).to eql("different_school")
-
-    fill_in :school_search, with: "Hampstead"
-    click_on "Search"
-
-    choose "Hampstead School"
-    click_on "Continue"
-
-    expect(claim.reload.current_school).to eql schools(:hampstead_school)
-
-    expect(page).to have_text(I18n.t("tslr.questions.mostly_teaching_eligible_subjects"))
-  end
-
-  scenario "chooses an ineligible school" do
-    claim = start_tslr_claim
-    choose_qts_year
-    choose_school schools(:hampstead_school)
-
-    expect(claim.reload.claim_school).to eq schools(:hampstead_school)
-    expect(page).to have_text("You’re not eligible")
-    expect(page).to have_text("Hampstead School is not an eligible school")
-  end
-
-  scenario "no longer teaching" do
-    claim = start_tslr_claim
-    choose_qts_year
-    choose_school schools(:penistone_grammar_school)
-
-    choose_still_teaching "No"
-
-    expect(claim.reload.employment_status).to eq("no_school")
-    expect(page).to have_text("You’re not eligible")
-    expect(page).to have_text("You must be still working as a teacher to be eligible")
-  end
-
-  scenario "did not teach at least half their time in an eligible subject" do
-    claim = start_tslr_claim
-    choose_qts_year
-    choose_school schools(:penistone_grammar_school)
-    choose_still_teaching
-
-    expect(page).to have_text(I18n.t("tslr.questions.mostly_teaching_eligible_subjects"))
-    choose "No"
-    click_on "Continue"
-
-    expect(claim.reload.mostly_teaching_eligible_subjects).to eq(false)
-    expect(page).to have_text("You’re not eligible")
-    expect(page).to have_text("You must have spent at least half your time teaching an eligible subject.")
-  end
-
   scenario "Teacher cannot go to mid-claim page before starting a claim" do
     visit claim_path("qts-year")
     expect(page).to have_current_path(root_path)
